requirements.txt
discord.py~=2.5.2
python-dotenv~=1.1.0
mcrcon~=0.7.0
asyncio~=3.4.3
pip~=25.0.1
attrs~=25.3.0
idna~=3.10
multidict~=6.4.4
propcache~=0.3.1
yarl~=1.20.0
aiohttp~=3.12.9
aiosignal~=1.3.2
frozenlist~=1.6.2
aiohappyeyeballs~=2.6.1
PyNaCl~=1.5.0
yt-dlp~=2025.6.9